+++ docs/changelog.md
@@ -15,6 +15,7 @@ ITables ChangeLog
 - We have improve the compatibility with dark themes ([#255](https://github.com/mwouts/itables/issues/255))
 - We now enforce non-sparse index when displaying Pandas Style objects with a multiindex ([#254](https://github.com/mwouts/itables/issues/254))
 - Export buttons are shown when using `show(df, buttons=...)` on a Pandas Style object ([#259](https://github.com/mwouts/itables/issues/259))
+- We have fixed a side effect when using `logging` ([#265](https://github.com/mwouts/itables/issues/265))
 
 
 2.0.0 (2024-03-16)